# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a modern RSS reader API.",
        "title": "Gist API",
        "contact": {},
        "version": "1.2.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/opml/import": {
            "post": {
                "description": "Start importing feeds and folders from an OPML file",
                "consumes": [
                    "multipart/form-data",
                    "text/xml"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "opml"
                ],
                "summary": "Import OPML",
                "parameters": [
                    {
                        "type": "file",
                        "description": "OPML file to import",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.importStartedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Cancel the current import task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "opml"
                ],
                "summary": "Cancel Import",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.importCancelledResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.importCancelledResponse": {
            "type": "object",
            "properties": {
                "cancelled": {
                    "type": "boolean"
                }
            }
        },
        "internal_handler.importStartedResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}